Added post saving using htmx

+++ apps/posts/views.py
@@ -65,16 +65,17 @@ def toggle_like(request, post_id):
 
 @login_required
 def toggle_save(request, post_id):
-    post = get_object_or_404(Post, pk=post_id)
-    save, created = Save.objects.get_or_create(user=request.user, post=post)
+    if request.method == 'POST':
+        post = get_object_or_404(Post, pk=post_id)
+        save, created = Save.objects.get_or_create(user=request.user, post=post)
 
-    if not created:
-        save.delete()
-        messages.success(request, 'Unsaved the post.')
-    else:
-        messages.success(request, 'Saved the post.')
+        if not created:
+            save.delete()
+            post.saved = False
+        else:
+            post.saved = True
 
-    return redirect(request.META.get('HTTP_REFERER'), 'feed:feed')
+        return render(request, 'posts/partials/save_button.html', {'post': post})
 
 
 @login_required
